+++ game_server/src/protocols/mod.rs
@@ -74,67 +74,67 @@ pub async fn route_packet(
         },
         Some(protocol) if *protocol == Protocol::Action as u8 => {
             let capacity = channel_tx.capacity();
-            server_state.tx_pc_client_gameplay.store(capacity, std::sync::atomic::Ordering::Relaxed);
+            server_state.tx_pc_client_gameplay.store( capacity as f32 as u16, std::sync::atomic::Ordering::Relaxed);
             movement_protocol::process_movement(socket, data, channel_tx).await;
         },
         Some(protocol) if *protocol == Protocol::Interaction as u8 => {
             let capacity = channel_map_tx.capacity();
-            server_state.tx_mc_client_gameplay.store(capacity, std::sync::atomic::Ordering::Relaxed);
+            server_state.tx_mc_client_gameplay.store(capacity as f32 as u16, std::sync::atomic::Ordering::Relaxed);
             interaction_protocol::process_interaction(socket, data, channel_map_tx).await;
         },
         Some(protocol) if *protocol == Protocol::Build as u8 => {
             let capacity = channel_map_tx.capacity();
-            server_state.tx_mc_client_gameplay.store(capacity, std::sync::atomic::Ordering::Relaxed);
+            server_state.tx_mc_client_gameplay.store(capacity as f32 as u16, std::sync::atomic::Ordering::Relaxed);
             build_protocol::process(socket, data, channel_map_tx).await;
         },
         Some(protocol) if *protocol == Protocol::TileAttacksWalker as u8 => { // used by mobs and towers.
             let capacity = channel_map_tx.capacity();
-            server_state.tx_mc_client_gameplay.store(capacity, std::sync::atomic::Ordering::Relaxed);
+            server_state.tx_mc_client_gameplay.store(capacity as f32 as u16, std::sync::atomic::Ordering::Relaxed);
             tile_attacks_walker_protocol::process(socket, data, channel_map_tx).await;
         },
         Some(protocol) if *protocol == Protocol::SpawnMob as u8 => {
             let capacity = channel_map_tx.capacity();
-            server_state.tx_mc_client_gameplay.store(capacity, std::sync::atomic::Ordering::Relaxed);
+            server_state.tx_mc_client_gameplay.store(capacity as f32 as u16, std::sync::atomic::Ordering::Relaxed);
             spawn_mob_protocol::process(socket, data, channel_map_tx).await;
         },
         Some(protocol) if *protocol == Protocol::MobMoves as u8 => {
             let capacity = channel_map_tx.capacity();
-            server_state.tx_mc_client_gameplay.store(capacity, std::sync::atomic::Ordering::Relaxed);
+            server_state.tx_mc_client_gameplay.store(capacity as f32 as u16, std::sync::atomic::Ordering::Relaxed);
             mob_moves_protocol::process(socket, data, channel_map_tx).await;
         },
         Some(protocol) if *protocol == Protocol::ControlMob as u8 => {
             let capacity = channel_map_tx.capacity();
-            server_state.tx_mc_client_gameplay.store(capacity, std::sync::atomic::Ordering::Relaxed);
+            server_state.tx_mc_client_gameplay.store(capacity as f32 as u16, std::sync::atomic::Ordering::Relaxed);
             claim_mob_ownership::process(socket, data, channel_map_tx).await;
         },
         Some(protocol) if *protocol == Protocol::AttackMob as u8 => {
             let capacity = channel_map_tx.capacity();
-            server_state.tx_mc_client_gameplay.store(capacity, std::sync::atomic::Ordering::Relaxed);
+            server_state.tx_mc_client_gameplay.store(capacity as f32 as u16, std::sync::atomic::Ordering::Relaxed);
             attack_mob_protocol::process(socket, data, channel_map_tx).await;
         },
         Some(protocol) if *protocol == Protocol::MissingPackets as u8 => {
             let capacity = channel_map_tx.capacity();
-            server_state.tx_mc_client_gameplay.store(capacity, std::sync::atomic::Ordering::Relaxed);
+            server_state.tx_mc_client_gameplay.store(capacity as f32 as u16, std::sync::atomic::Ordering::Relaxed);
             missing_packages_protocol::process_request(player_id, data, missing_packets);
         },
         Some(protocol) if *protocol == Protocol::AttackTower as u8 => {
             let capacity = channel_tower_tx.capacity();
-            server_state.tx_tc_client_gameplay.store(capacity, std::sync::atomic::Ordering::Relaxed);
+            server_state.tx_tc_client_gameplay.store(capacity as f32 as u16, std::sync::atomic::Ordering::Relaxed);
             attack_tower_protocol::process(socket, data, channel_tower_tx).await;
         },
         Some(protocol) if *protocol == Protocol::RepairTower as u8 => {
             let capacity = channel_tower_tx.capacity();
-            server_state.tx_tc_client_gameplay.store(capacity, std::sync::atomic::Ordering::Relaxed);
+            server_state.tx_tc_client_gameplay.store(capacity as f32 as u16, std::sync::atomic::Ordering::Relaxed);
             repair_tower_protocol::process(socket, data, channel_tower_tx).await;
         },
         Some(protocol) if *protocol == Protocol::RepairTower as u8 => {
             let capacity = channel_tower_tx.capacity();
-            server_state.tx_tc_client_gameplay.store(capacity, std::sync::atomic::Ordering::Relaxed);
+            server_state.tx_tc_client_gameplay.store(capacity as f32 as u16, std::sync::atomic::Ordering::Relaxed);
             repair_tower_protocol::process(socket, data, channel_tower_tx).await;
         },
         Some(protocol) if *protocol == Protocol::ChatMessage as u8 => {
             let capacity = channel_chat_tx.capacity();
-            server_state.tx_cc_client_gameplay.store(capacity, std::sync::atomic::Ordering::Relaxed);
+            server_state.tx_cc_client_gameplay.store(capacity as f32 as u16, std::sync::atomic::Ordering::Relaxed);
             chat_message_protocol::process(socket, data, channel_chat_tx).await;
         },
         unknown_protocol => {
